backend/requirements.txt
django~=2.1.5
djangorestframework~=3.9.1
django-cors-headers